Spliting MicrosoftGraph Extensions class into several classes
