+++ Microsoft.Toolkit.Uwp.Services/Services/MicrosoftGraph/Extensions/MicrosoftGraphUserExtensions.cs
@@ -14,18 +14,15 @@
 
 namespace Microsoft.Toolkit.Uwp.Services.MicrosoftGraph
 {
-    using System.Text;
     using System.Threading;
     using System.Threading.Tasks;
     using Graph;
 
     /// <summary>
-    /// GraphServiceClient Extensions
+    /// User's GraphServiceClient Extensions
     /// </summary>
-    public static class MicrosoftGraphExtensions
+    public static class MicrosoftGraphUserExtensions
     {
-     
-       
         /// <summary>
         /// Retrieve user's profile.
         /// </summary>
@@ -41,7 +38,7 @@ namespace Microsoft.Toolkit.Uwp.Services.MicrosoftGraph
                 return await graphClient.Me.Request().GetAsync(cancellationToken);
             }
 
-            string selectedProperties = selectFields.FormatSelectedFields();
+            string selectedProperties = MicrosoftGraphHelper.FormatString<MicrosoftGraphUserFields>(selectFields);
 
             return await graphClient.Me.Request().Select(selectedProperties).GetAsync(cancellationToken);
         }
@@ -92,76 +89,5 @@ namespace Microsoft.Toolkit.Uwp.Services.MicrosoftGraph
             return await graphClient.GetMePhotoAsync(CancellationToken.None);
         }
 
-        private const string OrderBy = "receivedDateTime desc";
-
-        /// <summary>
-        /// Retrieve the user's emails
-        /// </summary>
-        /// <param name="graphClient">Microsoft Graph Client's instance</param>
-        /// <param name="topMessages">The number of messages to return in a response</param>
-        /// <param name="selectFields">array of fields Microsoft Graph has to include in the response</param>
-        /// <returns>A strongly type including a list of the messages</returns>
-        public static async Task<IUserMessagesCollectionPage> GetUserMessagesAsync(this GraphServiceClient graphClient, int topMessages, MicrosoftGraphMessageFields[] selectFields)
-        {
-            return await graphClient.GetUserMessagesAsync(topMessages, selectFields, CancellationToken.None);
-        }
-
-        /// <summary>
-        /// Retrieve the user's emails
-        /// </summary>
-        /// <param name="graphClient">Microsoft Graph Client's instance</param>
-        /// <param name="topMessages">The number of messages to return in a response</param>
-        /// <param name="selectFields">array of fields Microsoft Graph has to include in the response</param>
-        /// <param name="cancellationToken">The <see cref="CancellationToken"/> for the request.</param>
-        /// <returns>A strongly type including a list of the messages</returns>
-        public static async Task<IUserMessagesCollectionPage> GetUserMessagesAsync(this GraphServiceClient graphClient, int topMessages, MicrosoftGraphMessageFields[] selectFields, CancellationToken cancellationToken)
-        {
-            if (selectFields == null)
-            {
-                return await graphClient.Me.Messages.Request().Top(topMessages).OrderBy(OrderBy).GetAsync(cancellationToken);
-            }
-
-            string selectedProperties = selectFields.FormatSelectedFields();
-
-            return await graphClient.Me.Messages.Request().Top(topMessages).OrderBy(OrderBy).Select(selectedProperties).GetAsync();
-        }
-
-        /// <summary>
-        /// Format an unique string with each array's items.
-        /// </summary>
-        /// <param name="selectFields">an array of MicrosoftGraphUserFields containing the fields</param>
-        /// <returns>a string with all fields separate by a comma.</returns>
-        private static string FormatSelectedFields(this MicrosoftGraphUserFields[] selectFields)
-        {
-            return FormatString<MicrosoftGraphUserFields>(selectFields);
-        }
-
-        /// <summary>
-        /// Format an unique string with each array's items.
-        /// </summary>
-        /// <param name="selectFields">an array of MicrosoftGraphUserFields containing the fields</param>
-        /// <returns>a string with all fields separate by a comma.</returns>
-        private static string FormatSelectedFields(this MicrosoftGraphMessageFields[] selectFields)
-        {
-            return FormatString<MicrosoftGraphMessageFields>(selectFields);
-        }
-
-        private static string FormatString<T>(T[] t)
-        {
-            StringBuilder sb = new StringBuilder();
-
-            foreach (var field in t)
-            {
-                sb.Append(field.ToString());
-                sb.Append(",");
-            }
-
-            string tempo = sb.ToString();
-
-            // Remove the trailing comma character
-            int lastPosition = tempo.Length - 1;
-
-            return tempo.Substring(0, lastPosition);
-        }
     }
 }
